before adding data channel scheduling function

+++ contiki/core/net/mac/prk-mcs/protocolsignaling.c
@@ -110,7 +110,7 @@ void printLinkERTable()
 /* update the link er table according to received ER information 
  * Note that these info are assumed to be valid. So caller should do validness checking.
 */
-void updateLinkER(uint8_t link_index, uint16_t er_version, int8_t I_edge)
+void updateLinkER(uint8_t link_index, uint8_t er_version, int8_t I_edge)
 {
 	uint8_t link_er_index = findLinkERTableIndex(link_index); 
 	if (link_er_index == INVALID_INDEX)
@@ -315,13 +315,14 @@ bool prepareERSegment(uint8_t *ptr)
 {
 	bool prepared = false;
 
+	//if (linkERTable[er_sending_index].primary || linkERTable[er_sending_index].secondary)
 	int8_t outbound_ed = getOutboundED(linkERTable[er_sending_index].receiver);
 	if (inER(outbound_ed, linkERTable[er_sending_index].I_edge))
 	{
 		memcpy(ptr, &(linkERTable[er_sending_index].link_index), sizeof(uint8_t));
 		ptr += sizeof(uint8_t);
-		memcpy(ptr, &(linkERTable[er_sending_index].er_version), sizeof(uint16_t));
-		ptr += sizeof(uint16_t);
+		memcpy(ptr, &(linkERTable[er_sending_index].er_version), sizeof(uint8_t));
+		ptr += sizeof(uint8_t);
 		memcpy(ptr, &(linkERTable[er_sending_index].I_edge), sizeof(int8_t));
 		ptr += sizeof(int8_t);
