+++ contiki/core/net/mac/prk-mcs/protocolsignaling.h
@@ -10,7 +10,7 @@ typedef struct {
 	/* index within all active links, used to compute priority */
 	uint8_t link_index;
 	/* ER version of the receiver */
-	uint16_t er_version;
+	uint8_t er_version;
 	int8_t I_edge;
 	/* each bit indicates the conflict relation with corresponding link in local er table */
 	uint16_t primary;
@@ -20,7 +20,7 @@ typedef struct {
 void protocolSignalingInit();
 uint8_t findLinkERTableIndex(uint8_t link_index);
 bool prepareERSegment(uint8_t *ptr);
-void updateLinkER(uint8_t link_index, uint16_t er_version, int8_t I_edge);
+void updateLinkER(uint8_t link_index, uint8_t er_version, int8_t I_edge);
 void updateConflictGraphForERChange(uint8_t link_er_index);
 void updateConflictGraphForLocalERChange(uint8_t local_link_er_index);
 #endif /* __PROTOCOLSIGNALING_H__ */
